fix import cycle and deep nesting sonarlint issues
