+++ app/services/gerichtsfinder/checkURLs.ts
@@ -55,28 +55,35 @@ function checkUrl(urlString: string) {
         },
       )
       .on("error", () => {
-        http
-          .get(normalizeURL(urlString, "http"), (response) => {
-            warnIfNot200(urlString, response);
-            resolve([urlString, response.responseUrl]);
-          })
-          .on("error", () => {
-            https
-              .get(
-                normalizeURL(urlString.replace("www.", ""), "https"),
-                (response) => {
-                  warnIfNot200(urlString, response);
-                  resolve([urlString, response.responseUrl]);
-                },
-              )
-              .on("error", () => {
-                if (urlString)
-                  console.error(`No valid redirect found for ${urlString}`);
-                resolve([urlString, ""]);
-              });
-          });
+        attemptHttp(urlString, resolve);
       }),
   );
+}
+
+function attemptHttp(
+  urlString: string,
+  resolve: (value: [string, string] | PromiseLike<[string, string]>) => void,
+) {
+  http
+    .get(normalizeURL(urlString, "http"), (response) => {
+      warnIfNot200(urlString, response);
+      resolve([urlString, response.responseUrl]);
+    })
+    .on("error", () => {
+      https
+        .get(
+          normalizeURL(urlString.replace("www.", ""), "https"),
+          (response) => {
+            warnIfNot200(urlString, response);
+            resolve([urlString, response.responseUrl]);
+          },
+        )
+        .on("error", () => {
+          if (urlString)
+            console.error(`No valid redirect found for ${urlString}`);
+          resolve([urlString, ""]);
+        });
+    });
 }
 
 async function checkAllURLS() {
